migrate utils from CustomerFlow/StoreOwnerFlow to customer, merchant and shopper
