+++ tests/e2e/utils/src/flows/index.js
@@ -1,0 +1,16 @@
+/**
+ * Internal dependencies
+ */
+const flowConstants = require( './constants' );
+const flowExpressions = require( './expressions' );
+const customer = require( './customer' );
+const merchant = require( './merchant' );
+const shopper = require( './shopper' );
+
+module.exports = {
+	...flowConstants,
+	...flowExpressions,
+	customer,
+	merchant,
+	shopper,
+};
