+++ tests/e2e/utils/src/old-flows.js
@@ -1,0 +1,18 @@
+/**
+ * Internal dependencies
+ */
+const {
+	customer,
+	merchant,
+	shopper
+} = require( './flows' );
+
+const CustomerFlow = {
+	...shopper,
+	...customer,
+};
+
+export {
+	CustomerFlow,
+	merchant as StoreOwnerFlow,
+};
